requirements.txt
screeninfo==0.8.1
watchdog==3.0.0
Pillow==10.2.0
